refactor: modulized files and request
- gather db request to one folter
- fix trivial error
