+++ src/app/action.ts
@@ -1,7 +1,7 @@
 "use server";
 
 import { redirect } from "next/navigation";
-import { setUserName } from "./util/server";
+import { setUserName } from "../util/server";
 
 export async function action(data: FormData) {
   const { userName } = Object.fromEntries(data);
