+++ src/app/page.tsx
@@ -1,13 +1,8 @@
-import { cookies } from "next/headers";
-import { redirect } from "next/navigation";
 import Button from "@/component/Button";
 import { action } from "./action";
 import styles from "./page.module.css";
 
 export default function Home() {
-  if (cookies().get("userName")) {
-    redirect("/menu");
-  }
   return (
     <main className={styles.main}>
       <form action={action} className={styles.description}>
