+++ src/crawling/index.ts
@@ -1,36 +1,5 @@
 import { CheerioAPI, load } from "cheerio";
-
-const ingredientList = [
-  { name: "calories", className: "bg1" },
-  { name: "saturatedFat", className: "bg2" },
-  { name: "sodium", className: "bg3" },
-  { name: "carbohydrate", className: "bg4" },
-  { name: "sugars", className: "bg5" },
-  { name: "caffeine", className: "bg6" },
-  { name: "protain", className: "bg7" },
-] as const;
-
-export const initialIngredient = ingredientList.reduce(
-  (acc, ingredient) => ({ ...acc, [ingredient.name]: 0 }),
-  {}
-) as Readonly<Record<IngredientKey, number>>;
-
-type IngredientKey = (typeof ingredientList)[number]["name"];
-
-export type MenuContentsProps = {
-  title: string;
-  list: MenuProps[];
-};
-
-export type MenuProps = {
-  photo: string;
-  name: {
-    kor: string;
-    eng: string;
-  };
-  description: string;
-  info: typeof initialIngredient;
-};
+import { MenuContentsProps, ingredientList, initialIngredient } from "@/type";
 
 export async function getMenu(url: string) {
   const urlObject = new URL(url);
